+++ src/components/modals/ComponentDetailModal/styles.ts
